Restore tile padding
Aim is to resolve
* #5

This is quite fiddly and awkward, because it needs to support multiple outputs and potential size reduction for compatibility with 'default' StarDist models. This means we can't simply wrap a DnnModel in an ImageOp (as in the original implementation) and have it take care of any padding.